+++ udata/core/dataset/api.py
@@ -221,7 +221,7 @@ class DatasetApiParser(ModelApiParser):
             except Topic.DoesNotExist:
                 pass
             else:
-                datasets = datasets.filter(id__in=[d.id for d in topic.datasets])
+                datasets = datasets.filter(id__in=topic.get_nested_elements_ids("Dataset"))
         if args.get("dataservice"):
             if not ObjectId.is_valid(args["dataservice"]):
                 api.abort(400, "Dataservice arg must be an identifier")